requirements.txt
flask
flask_cors
pymongo
pandas
python-dotenv
